LayoutTests/platform/wpe/imported/w3c/web-platform-tests/editing/run/caret-navigation-around-line-break-expected.txt

line1line2
line1
line2
line1line2
line1line2
line1line2

PASS no separator - move down
PASS no separator - move up
PASS no separator - move right
PASS no separator - move left
PASS <br> separator - move down
PASS <br> separator - move up
PASS <br> separator - move right
PASS <br> separator - move left
FAIL <wbr> separator - move down assert_equals: Caret moved to line 2 expected Text node "line2" but got Text node "line1"
FAIL <wbr> separator - move up assert_equals: Caret is in line 2 expected Text node "line2" but got Text node "line1"
PASS <wbr> separator - move right
FAIL <wbr> separator - move left assert_equals: Caret is in line 2 expected Text node "line2" but got Text node "line1"
FAIL <span> separator - move down assert_equals: Caret moved to line 2 expected Text node "line2" but got Text node "line1"
FAIL <span> separator - move up assert_equals: Caret is in line 2 expected Text node "line2" but got Text node "line1"
PASS <span> separator - move right
FAIL <span> separator - move left assert_equals: Caret is in line 2 expected Text node "line2" but got Text node "line1"
FAIL two <span> separators - move down assert_equals: Caret moved to line 2 expected Text node "line2" but got Text node "line1"
FAIL two <span> separators - move up assert_equals: Caret is in line 2 expected Text node "line2" but got Text node "line1"
PASS two <span> separators - move right
FAIL two <span> separators - move left assert_equals: Caret is in line 2 expected Text node "line2" but got Text node "line1"
